Issue #1321: Fix rename pyimod04_importers to pyimod03_importers.
