+++ PyInstaller/loader/pyiboot01_bootstrap.py
@@ -12,10 +12,10 @@
 # Only python built-in modules can be used.
 
 import sys
-import pyimod04_importers
+import pyimod03_importers
 
 # Extend Python import machinery by adding PEP302 importers to sys.meta_path.
-pyimod04_importers.install()
+pyimod03_importers.install()
 
 
 ### Bootstrap process is complete.
