+++ PyInstaller/loader/pyimod02_archive.py
@@ -231,7 +231,7 @@ class Cipher(object):
         """
         # Not-so-easy way: at bootstrap time we have to load the module from the
         # temporary directory in a manner similar to pyi_importers.CExtensionImporter.
-        from pyimod04_importers import CExtensionImporter
+        from pyimod03_importers import CExtensionImporter
         importer = CExtensionImporter()
         # NOTE: We _must_ call find_module first.
         # The _AES.so module exists only in PyCrypto 2.6 and later. Try to import
