+++ PyInstaller/loader/rthooks/pyi_rth_PIL_Image.py
@@ -9,7 +9,7 @@
 
 
 import sys
-import pyimod04_importers as pyi_importers
+import pyimod03_importers as pyi_importers
 
 
 # Find FrozenImporter object from sys.meta_path.
